+++ src/wordviz/plotting3d.py
@@ -236,9 +236,10 @@ class Visualizer3D(BaseVisualizer):
         return fig
 
 
-    def plot_clusters(self, n_clusters=5, method='kmeans', red_method='auto', show_centers=False, grid=True, theme='light1', title=None, nlabels=0, use_subset=False):
+    def plot_clusters(self, n_clusters: int = 5, method: str = 'kmeans', metric: str = None, red_method: str = 'auto', show_centers: bool = False, grid: bool = True, theme: str = 'light1', title: str = None, nlabels: int = 0, use_subset: bool = False) -> go.Figure:
         '''
         Creates a 3D scatterplot of clustered embeddings using a clustering algorithm.
+        Dimensionality reduction is performed before clustering, in order to enhance visualization and reduce noise.
 
         Parameters:
         -----------
@@ -246,6 +247,8 @@ class Visualizer3D(BaseVisualizer):
             Number of clusters to generate.
         method : str, default='kmeans'
             Clustering method to use ('kmeans' or others supported by create_clusters).
+        metric : str, optional
+            Distance metric to use for clustering.
         red_method : str, default='auto' -> auto is deprecated, will default to pca in future releases
             Dimensionality reduction method to apply ('pca', 'tsne', 'umap', etc.). If 'auto' searches for cached reduction, if None runs pca.
         show_centers : bool, default=False
@@ -275,7 +278,7 @@ class Visualizer3D(BaseVisualizer):
         )
         reduced_emb, tokens = self._set_embeddings(use_subset=use_subset, red_method=red_method, dims=3)
 
-        clusters, centers, reduced_emb = create_clusters(reduced_emb, n_clusters=n_clusters, method=method)
+        clusters, centers, reduced_emb = create_clusters(reduced_emb, n_clusters=n_clusters, method=method, metric=metric)
 
         fig = self._setup_3d(reduced_emb=reduced_emb, theme=theme, grid=grid, tokens=tokens, title=title, def_title=f"3D Clustering Scatterplot", labels=clusters, for_clustering=True)